+++ mitra-config/src/config.rs
@@ -128,6 +128,7 @@ impl Config {
                 matches!(self.environment, Environment::Development),
             fetcher_timeout: self.federation.fetcher_timeout,
             deliverer_timeout: self.federation.deliverer_timeout,
+            deliverer_log_response_length: self.federation.deliverer_log_response_length,
             fep_8b32_eddsa_enabled: self.federation.fep_8b32_eddsa_enabled,
         }
     }
@@ -190,6 +191,7 @@ pub struct Instance {
     pub is_private: bool,
     pub fetcher_timeout: u64,
     pub deliverer_timeout: u64,
+    pub deliverer_log_response_length: usize,
 
     pub fep_8b32_eddsa_enabled: bool,
 }
@@ -229,6 +231,7 @@ impl Instance {
             is_private: true,
             fetcher_timeout: 0,
             deliverer_timeout: 0,
+            deliverer_log_response_length: 0,
             fep_8b32_eddsa_enabled: false,
         }
     }
@@ -252,6 +255,7 @@ mod tests {
             is_private: true,
             fetcher_timeout: 0,
             deliverer_timeout: 0,
+            deliverer_log_response_length: 0,
             fep_8b32_eddsa_enabled: false,
         };
 
@@ -276,6 +280,7 @@ mod tests {
             is_private: true,
             fetcher_timeout: 0,
             deliverer_timeout: 0,
+            deliverer_log_response_length: 0,
             fep_8b32_eddsa_enabled: false,
         };