+++ mitra-config/src/federation.rs
@@ -3,6 +3,7 @@ use serde::Deserialize;
 const fn default_federation_enabled() -> bool { true }
 const fn default_fetcher_timeout() -> u64 { 300 }
 const fn default_deliverer_timeout() -> u64 { 30 }
+const fn default_deliverer_log_response_length() -> usize { 75 }
 const fn default_fep_e232_enabled() -> bool { false }
 const fn default_fep_8b32_eddsa_enabled() -> bool { false }
 
@@ -15,6 +16,9 @@ pub struct FederationConfig {
     pub(super) fetcher_timeout: u64,
     #[serde(default = "default_deliverer_timeout")]
     pub(super) deliverer_timeout: u64,
+    #[serde(default = "default_deliverer_log_response_length")]
+    pub(super) deliverer_log_response_length: usize,
+
     pub(super) proxy_url: Option<String>,
     pub(super) onion_proxy_url: Option<String>,
     pub(super) i2p_proxy_url: Option<String>,
@@ -31,6 +35,7 @@ impl Default for FederationConfig {
             enabled: default_federation_enabled(),
             fetcher_timeout: default_fetcher_timeout(),
             deliverer_timeout: default_deliverer_timeout(),
+            deliverer_log_response_length: default_deliverer_log_response_length(),
             proxy_url: None,
             onion_proxy_url: None,
             i2p_proxy_url: None,