+++ src/activitypub/agent.rs
@@ -37,6 +37,7 @@ pub fn build_federation_agent(
         is_instance_private: instance.is_private,
         fetcher_timeout: instance.fetcher_timeout,
         deliverer_timeout: instance.deliverer_timeout,
+        deliverer_log_response_length: instance.deliverer_log_response_length,
         proxy_url: instance.proxy_url.clone(),
         onion_proxy_url: instance.onion_proxy_url.clone(),
         i2p_proxy_url: instance.i2p_proxy_url.clone(),