+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Support changing logger configuration with `RUST_LOG` environment variable.
+
 ## [2.4.1] - 2023-11-23
 
 ### Added