+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Support changing logger configuration with `RUST_LOG` environment variable.
 
+### Changed
+
+- Allow `assertionMethod` property value to be object.
+
 ## [2.4.1] - 2023-11-23
 
 ### Added